Fix LzBlock
Remove TODO items.
Previously if the MRU list got too big we'd fail an assertion.
The new rule says that the maximum MRU size is flexible.

The rule saying that the size was fixed goes back to LzStream.C.
That file format did not have an explicit delete command.
When the MRU list was full, we had to delete old data.