Dump more status in LzBlock.C to stdout.
Each block now prints something like this:
count   bytes   bts per reason
15477   16858.4 8.71406 Index
11542   1426.36 0.988637        Delete
3949    1410.79 2.85802 Write

Write takes almost 3 bits per.
It's a Boolean, so it should never be worse than 1 bit per.
At a bare minimum we could hard code that to 50/50 to improve it.

Delete takes just slightly less than 1 bit per.
This could be a lot better, too.

Index contains the bulk of the data, around 85%.
The input is more complicated.
We'll need to dump more statistics, or try something different, to know if the current algorithm is working.