Test an alternate way to save compress parts of the data.
There are a lot of decisions to save or not to save an item.
We have been using a running total from the beginning of the file.
We have different totals for each string length.

This commit adds an option where we use statistics for the current block.
We use read ahead to get statics for the entire block.
Potentially this data is more relevant than the data from the beginning of the file.
We are keeping stats for length = 2, length = 3, and length > 3.
We have less precision for this case because we have to write the statistics to the file explicitly.

This new attempt never goes to the output file.
Only to the log on std::cerr.

This new attempt gave results similar to the previous version.
Sometimes better, sometimes worse.