Explore a different way to store the index data.
Currently we are using counters to track the data in real time.
And those counters set the probabilities for each index we write.

The new test data asks what would happen if we used a different strategy:
  Count all of the probabilities for the block,
  Store a brief summary of that in the block header.

The new strategy seems to save a few bytes each block.
If it was just the saved bytes, this change would not be so interesting.
The new strategy will make it faster to encode and decode the data.
It takes work to update the probabilities.
Now we are only doing that once per block, not once per index.