Bug fix
We were throwing away a lot of data and leaving a lot of empty space
in the table.
We were always overwriting the first index for each hash.
Now we are doing a round robin,
overwriting the oldest entry if we run out of space.

Add detailed comments about bugs, and other TODOs.

Add *a lot* of logging.
This found the bug that was fixed in this commit.
And inspired a lot of the new comments.