Bug fix
The one byte context data was not being stored right.
The order of the two bytes was reversed.
This caused us to miss a lot of matches.
This bug fix removes a TODO.
Added some logging to see that we are storing the bytes correctly now.

Notice new TODO items.