OneByteRange now handles overflow correctly.
As planned, we now walk through every entry that starts with the same byte and divide it's frequency by 2.
If the frequency goes to 0, delete that entry.

Code refactoring:
The Range class helps iterate over all entries that have a specific byte of context.
This code was used to measure the frequency of each possible byte.
Now it is also used to handle the overflow situation.

Note:  Overflows did not happen as often as the logs would have suggested.
When the previous code saw an overflow, it would print something to the log, but it did not change anything.
So the next time we saw the same byte of context, we'd print the same messages.
The new code still reports each time it detects an overflow.
But it reports a lot less often because we are fixing the overflow.

As expected, this did not affect any test cases unless we saw an overflow in that case.

When there was an overflow, I saw mixed results.
Two files got a little smaller.
The really big and redundant log file got a tiny bit bigger.