Add an option to save more strings.
Bug fix:
The old way we were computing the cost of the hashed data was way low.

Added a more realistic way to save the hashed data.
We now keep track of which indexes are used more frequently.
So we can compress these index values.

Add a new version of the compression routine.
Previously each run would save strings of a specific size.
Now we can save strings in a range of sizes all in one table.