Bug fix:  Fix a segmentation violation.
Remove a segmentation violation in the code.
There was an error in the logic in a for loop.

Rename the test files.
So VS Code won't try to syntax check them.

Fix the build process.
Running the debugger will now automatically rebuild.
And the build will now work no matter which file tab is selected.